Brings the staged RPCS3 upstream merges v0.0.38 through v0.0.42 with the
Android/ARM64 work built on top of them:

- LLVM 22.1.8 built from source for Android (persistent build cache,
  MinGW host toolchain for the NATIVE tblgen build, per-ABI host triple)
- ARM64 fixes: SPU UABA checksum collision, cntvct_el0 decrementer
- v0.0.41 boot hang root cause: glslang GL_EXT_uniform_buffer_unsized_array
  gate with concrete UBO array bounds for Adreno
- Silent audio fix: cubeb default-device fall-through (AAudio cannot
  enumerate devices)
- Async shader compilation: GLSL->SPIR-V moved off the RSX thread
- ZCULL on Adreno: Turnip/Qualcomm driver_vendor entries, WFE-parked
  interruptible result wait, RSX<->PPU deadlock fix, rescue flush

+++ kernel/cellos/include/cellos/sys_ppu_thread.h
@@ -58,7 +58,7 @@ error_code sys_ppu_thread_stop(ppu_thread &ppu, u32 thread_id);
 error_code sys_ppu_thread_restart(ppu_thread &ppu);
 error_code _sys_ppu_thread_create(ppu_thread &ppu, vm::ptr<u64> thread_id,
                                   vm::ptr<ppu_thread_param_t> param, u64 arg,
-                                  u64 arg4, s32 prio, u32 stacksize, u64 flags,
+                                  u64 arg4, s32 prio, u64 stacksize, u64 flags,
                                   vm::cptr<char> threadname);
 error_code sys_ppu_thread_start(ppu_thread &ppu, u32 thread_id);
 error_code sys_ppu_thread_rename(ppu_thread &ppu, u32 thread_id,
